+++ file-kache-okio/src/commonMain/kotlin/com/mayakapps/kache/journal/Consts.kt
@@ -1,0 +1,8 @@
+package com.mayakapps.kache.journal
+
+internal const val JOURNAL_MAGIC = "JOURNAL"
+internal const val JOURNAL_VERSION: Byte = 1
+
+internal const val JOURNAL_FILE = "journal"
+internal const val JOURNAL_FILE_TEMP = "$JOURNAL_FILE.tmp"
+internal const val JOURNAL_FILE_BACKUP = "$JOURNAL_FILE.bkp"
