--- file-kache-okio/src/commonMain/kotlin/com/mayakapps/lrucache/FileSystem.kt
@@ -1,23 +1,0 @@
-package com.mayakapps.lrucache
-
-import okio.Closeable
-import okio.FileSystem
-import okio.Path
-import okio.use
-
-internal expect fun getSystemFileSystem(): FileSystem
-
-fun FileSystem.deleteContents(dir: Path) {
-    list(dir).forEach { deleteRecursively(it) }
-}
-
-fun FileSystem.atomicMove(source: Path, target: Path, deleteTarget: Boolean) {
-    if (deleteTarget) delete(target)
-    atomicMove(source, target)
-}
-
-inline fun <T : Closeable?, R> T.nullableUse(block: (T) -> R?): R? = try {
-    use(block)
-} catch (ex: NullPointerException) {
-    null
-}
