--- file-kache-okio/src/commonMain/kotlin/com/mayakapps/lrucache/journal/Consts.kt
@@ -1,8 +1,0 @@
-package com.mayakapps.lrucache.journal
-
-internal const val JOURNAL_MAGIC = "JOURNAL"
-internal const val JOURNAL_VERSION: Byte = 1
-
-internal const val JOURNAL_FILE = "journal"
-internal const val JOURNAL_FILE_TEMP = "$JOURNAL_FILE.tmp"
-internal const val JOURNAL_FILE_BACKUP = "$JOURNAL_FILE.bkp"
