--- file-kache-okio/src/commonMain/kotlin/com/mayakapps/lrucache/journal/Exceptions.kt
@@ -1,9 +1,0 @@
-package com.mayakapps.lrucache.journal
-
-import okio.IOException
-
-internal sealed class JournalException(override val message: String? = null) : IOException(message)
-
-internal class JournalInvalidHeaderException(override val message: String? = null) : JournalException(message)
-
-internal class JournalInvalidOpcodeException : JournalException()
