+++ file-kache-okio/src/jsMain/kotlin/com/mayakapps/kache/FileSystem.kt
@@ -1,0 +1,6 @@
+package com.mayakapps.kache
+
+import okio.FileSystem
+import okio.NodeJsFileSystem
+
+internal actual fun getSystemFileSystem(): FileSystem = NodeJsFileSystem
