--- file-kache-okio/src/jsMain/kotlin/com/mayakapps/lrucache/FileSystem.kt
@@ -1,6 +1,0 @@
-package com.mayakapps.lrucache
-
-import okio.FileSystem
-import okio.NodeJsFileSystem
-
-internal actual fun getSystemFileSystem(): FileSystem = NodeJsFileSystem
