+++ file-kache-okio/src/jvmMain/kotlin/com/mayakapps/kache/FileSystem.kt
@@ -1,0 +1,7 @@
+@file:JvmName("FileSystemJvmKt")
+
+package com.mayakapps.kache
+
+import okio.FileSystem
+
+internal actual fun getSystemFileSystem(): FileSystem = FileSystem.SYSTEM
