--- file-kache-okio/src/jvmMain/kotlin/com/mayakapps/lrucache/FileSystem.kt
@@ -1,7 +1,0 @@
-@file:JvmName("FileSystemJvmKt")
-
-package com.mayakapps.lrucache
-
-import okio.FileSystem
-
-internal actual fun getSystemFileSystem(): FileSystem = FileSystem.SYSTEM
