+++ file-kache-okio/src/nativeMain/kotlin/com/mayakapps/kache/FileSystem.kt
@@ -1,0 +1,5 @@
+package com.mayakapps.kache
+
+import okio.FileSystem
+
+internal actual fun getSystemFileSystem(): FileSystem = FileSystem.SYSTEM
