--- file-kache-okio/src/nativeMain/kotlin/com/mayakapps/lrucache/FileSystem.kt
@@ -1,5 +1,0 @@
-package com.mayakapps.lrucache
-
-import okio.FileSystem
-
-internal actual fun getSystemFileSystem(): FileSystem = FileSystem.SYSTEM
