+++ kache/src/commonMain/kotlin/com/mayakapps/kache/ConcurrentMutableMap.kt
@@ -1,0 +1,3 @@
+package com.mayakapps.kache
+
+internal expect class ConcurrentMutableMap<K : Any, V : Any>() : MutableMap<K, V>
