--- kache/src/commonMain/kotlin/com/mayakapps/lrucache/ConcurrentMutableMap.kt
@@ -1,3 +1,0 @@
-package com.mayakapps.lrucache
-
-internal expect class ConcurrentMutableMap<K : Any, V : Any>() : MutableMap<K, V>
