+++ kache/src/nonJvmMain/kotlin/com/mayakapps/kache/ConcurrentMutableMap.kt
@@ -1,0 +1,6 @@
+package com.mayakapps.kache
+
+import co.touchlab.stately.collections.IsoMutableMap
+
+internal actual class ConcurrentMutableMap<K : Any, V : Any> actual constructor() : MutableMap<K, V>,
+    IsoMutableMap<K, V>()
