--- kache/src/nonJvmMain/kotlin/com/mayakapps/lrucache/ConcurrentMutableMap.kt
@@ -1,6 +1,0 @@
-package com.mayakapps.lrucache
-
-import co.touchlab.stately.collections.IsoMutableMap
-
-internal actual class ConcurrentMutableMap<K : Any, V : Any> actual constructor() : MutableMap<K, V>,
-    IsoMutableMap<K, V>()
